Upgrade advanced trigger counter to be 32bit @ LA mode

+++ libsigrok4DSL/hardware/DSL/dsl.h
@@ -229,9 +229,9 @@ struct DSL_setting {
     //uint32_t trig_edge3_header;              // 27
     //uint16_t trig_edge3[NUM_TRIGGER_STAGES];
     uint32_t trig_count0_header;             // 28
-    uint16_t trig_count0[NUM_TRIGGER_STAGES];
+    uint32_t trig_count0[NUM_TRIGGER_STAGES];
     uint32_t trig_count1_header;             // 29
-    uint16_t trig_count1[NUM_TRIGGER_STAGES];
+    uint32_t trig_count1[NUM_TRIGGER_STAGES];
     //uint32_t trig_count2_header;             // 30
     //uint16_t trig_count2[NUM_TRIGGER_STAGES];
     //uint32_t trig_count3_header;             // 31
